+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - match zoom direction and step size on scrollwheel with eog
 - support pinch zoom [earboxer]
+- background switcher
 
 ## 2.3, started 20/10/21